Tests Written

+++ src/Controllers/Chats/chats_controllers.ts
@@ -1,14 +1,24 @@
-import { Controller, Post, Req, Res } from '@nestjs/common';
+import { Controller, Inject, Post, Req, Res } from '@nestjs/common';
 import express from 'express';
 import { ControllerPaths } from '../../Constants/contoller_paths';
 import { ChatMessageForUserService } from './Services/send_chat_message_service';
 import { ChatRoutes } from '../../Constants/route_paths';
 import { ThrowError } from '../../Errors/throw_error';
 import { types } from 'cassandra-driver';
+import { InjectionTokens } from '../../Constants/injection_tokens';
+import { CreateQueue } from '../../Queues/base';
+import { QueueNames } from '../../Constants/queues_redis';
+import { ConflictError, ConflictErrorTypes } from '../../Errors/conflict_error';
 
 @Controller(ControllerPaths.CHATS_CONTROLLER)
 export class ChatsController {
-  constructor(private readonly chatMessageService: ChatMessageForUserService) {}
+  private createQueue: CreateQueue;
+  constructor(
+    private readonly chatMessageService: ChatMessageForUserService,
+    @Inject(InjectionTokens.CreateQueue) createQueue: CreateQueue,
+  ) {
+    this.createQueue = createQueue;
+  }
 
   @Post(ChatRoutes.SEND_CHAT_MESSAGE)
   async sendChatMessage(
@@ -47,6 +57,9 @@ export class ChatsController {
   ) {
     try {
       const { delete_message_model } = req.body;
+      if (delete_message_model.length > 50) {
+        throw new ConflictError(ConflictErrorTypes.BATCH_LIMIT_EXCEED);
+      }
       await this.chatMessageService.deleteChatMessageForMe(
         delete_message_model,
       );
@@ -81,11 +94,15 @@ export class ChatsController {
         update_status_of_chat_message_model,
         updated_status,
       } = req.body;
-      await this.chatMessageService.updateStatusOfChatMessages(
+      if (update_status_of_chat_message_model.length > 50) {
+        throw new ConflictError(ConflictErrorTypes.BATCH_LIMIT_EXCEED);
+      }
+      this.chatMessageService.updateStatusOfChatMessages(
         reciever_id,
         update_status_of_chat_message_model,
         updated_status,
       );
+      return res.status(200).json(true);
     } catch (error) {
       throw new ThrowError(error, res);
     }
